Actually giving up on API. Explore claims data
The claims data is small enough to download in full, and the policies data API times out before you can actually do anything useful with it. Working on getting the policies JSON dump, but regardless, querying the API is unlikely to be helpful going forward.

Select the claims for single family homes that had building coverage and a positive payout. Will need to handle re-issued checks as well.